[REFACTOR]폴더명 수정
