+++ app/src/views/home/signup.html
@@ -8,10 +8,12 @@
     <title>Sign Up</title>
 
     <!-- Font Icon -->
-    <link rel="stylesheet" href="fonts/material-icon/css/material-design-iconic-font.min.css">
+
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/css/bootstrap.min.css" rel="stylesheet"
         integrity="sha384-KK94CHFLLe+nY2dmCWGMq91rCGa5gtU4mk92HdvYe+M/SXH301p5ILy+dN9+nJOZ" crossorigin="anonymous">
-    <!-- Main css -->
+   
+    
+
     <link rel="stylesheet" href="/css/home/signup.css">
 </head>
 
@@ -21,14 +23,14 @@
         <div class="container">
             <div class="signup-content">
                 <div class="signup-form">
-                    <form action ="/auth/register" method="POST" class="register-form" id="register-form">
+                    <form action ="/register" method="POST" class="register-form" id="register-form">
                         <h2>UniUnity 회원가입</h2>
                         <label for="univ_info" class="form-label">대학교</label>
-                        <select class="form-select" aria-label="univ_select" >
-                            <option selected>대학교 선택</option>
-                            <option value="1">One</option>
-                            <option value="2">Two</option>
-                            <option value="3">Three</option>
+                        <select class="selectpicker" aria-label="univ_select" id="'univ_select">
+                            <option selected>대학교 선택 ▼ </option>
+                            <option value="1">성신여자대학교</option>
+                            <option value="2">건국대학교</option>
+                            <option value="3">성균관대학교</option>
                           </select>
 
                           <label for="univ_info" class="form-label">유저 타입</label>
@@ -101,8 +103,7 @@
     </div>
 
     <!-- JS -->
-    <script src="vendor/jquery/jquery.min.js"></script>
-    <script src="js/signup.js"></script>
+    <script src="js/home/signup.js"></script>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha3/dist/js/bootstrap.bundle.min.js"
         integrity="sha384-ENjdO4Dr2bkBIFxQpeoTz1HIcje39Wm4jDKdf19U8gI4ddQ3GYNS7NTKfAdVQSZe"
         crossorigin="anonymous"></script>
